breaking: remove xboxButton, remove fullName, change name to be more like fullName, improve example scene for 3+ controllers, improve debugger, updater child of Hinput Settings

+++ Assets/Hinput/Scripts/Utils/Utils.cs
@@ -60,7 +60,7 @@ namespace HinputClasses.Internal {
 
 		// Returns true if stick is currently within a (Settings.stickType) degree arc oriented at angle.
 		public static bool PushedTowards (this Stick stick, float angle) { 
-			return (Mathf.Abs(Mathf.DeltaAngle(angle, stick.angle)) < ((int)Settings.stickType) / 2); 
+			return (Mathf.Abs(Mathf.DeltaAngle(angle, stick.angle)) < ((float)Settings.stickType) / 2); 
 		}
